feat(connect)!: set legacyGet header to false on cache.get and data.get #531
