+++ packages/connect/deno/tests/data.ts
@@ -39,7 +39,7 @@ test("data.get", async () => {
     assertEquals(h.service, "data");
     assertEquals(h.method, "GET");
     assertEquals(h.resource, "game-1");
-    assertEquals(h.headers?.get(HYPER_LEGACY_GET_HEADER), "true");
+    assertEquals(h.headers?.get(HYPER_LEGACY_GET_HEADER), "false");
     return Promise.resolve(new Request("http://localhost"));
   };
